Switched over to new Windows pthreads library
